deep_geo_backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deep-geo-backend"
version = "2.0.0"
description = "Moteur Deep GEO — Places API, crawl Markdown, audit RAG (Groq), Celery+Redis+PostgreSQL"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.9",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.2.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "groq>=0.11.0",
    "orjson>=3.9.0",
    "greenlet>=3.0.0",
    "trafilatura>=1.9.0",
    "honcho>=1.1.0",
]

[project.optional-dependencies]
# Crawl4AI tire Playwright — optionnel. Sans : fallback httpx+trafilatura.
crawl4ai = ["crawl4ai>=0.4.0"]
dev = ["ruff>=0.4.0", "mypy>=1.10.0"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"
